requirements.txt
pdfplumber
pandas
SQLAlchemy
psycopg2-binary
PyYAML
